+++ docs/RELEASES.md
@@ -16,7 +16,7 @@ Legend:
 
 ## <a name="7.4.0.0"></a>7.4.0.0 :whale: <sup>`6`</sup>
 - [ ] Hide closed accounts if there are no closed accounts
-- [ ] `210901>......` When there are accounts, show "create another account" (and category)
+- [x] `210901>210911` When there are accounts, show "create another account" (and category)
 - [x] `210831>210831` Change "tudo bem" by "saudável" in portuguese translation
 - [x] `210831>210831` Warning of "just numbers" in just numbers fields
 - [x] `210829>210831` Remove account url field (calculate no diacritics - take care equal url)